+++ lib/liveness_flutter.dart
@@ -3,6 +3,7 @@ library liveness_flutter;
 
 export 'src/camera/frame_quality.dart' show FrameQuality;
 export 'src/camera/liveness_capabilities.dart';
+export 'src/detection/flash_challenge.dart' show FlashChallenge;
 export 'src/detection/spoof_guard.dart';
 export 'src/models/models.dart';
 export 'src/theme/liveness_theme.dart';